[변경] RacingController 에서 InputView 사용하지 않게 변경 - PlayRacingController : 값 유효성 체크. 여기서만 InputView 사용. - RacingController : 값 초기화 - Util에 횟수 범위 확인하는 함수 추가
